requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
diskcache==5.6.3
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
llama_cpp_python==0.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.23.5
packaging==24.1
pillow==11.0.0
PyPDF2==3.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.2.2
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
